requirements.txt
requests
pydantic>=2.0
python-dotenv
tenacity
PyYAML
rich
